requirements.txt
alembic==1.4.3
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.2
numpy==1.19.2
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
torch==1.6.0
torchvision==0.7.0
Werkzeug==1.0.1
wrapt==1.12.1
